Refactor GetUnassignedIps to Helpers and add documentation
